+++ ocdsextensionregistry/__init__.py
@@ -1,3 +1,5 @@
+from .codelist import Codelist  # noqa: F401
+from .codelist_code import CodelistCode  # noqa: F401
 from .extension import Extension  # noqa: F401
 from .extension_version import ExtensionVersion  # noqa: F401
 from .extension_registry import ExtensionRegistry  # noqa: F401